requirements/prod.txt
# Basic
dataclasses-json==0.5.2

#Flask
flask==1.1.2
flask-login==0.5.0
flask-session==0.3.2
gunicorn==20.0.4
pyjwt==1.7.1

# Javascript
Flask-Assets==2.0

# flask-wtf
flask-wtf==0.14.3
email-validator==1.1.1

# Datasbase
flask-sqlalchemy==2.4.4
flask-migrate==2.5.3

# Socket IO
flask-socketio==4.3.1
gevent-websocket==0.10.1